chore(auth-controller): move sanitization to middleware
I moved the auth data sanitization into a middleware so as to
adhere to the single responsibility principle
